1.split XStreamFactory class 2.all "handler" fields changed type to String-> Handler
