+++ src/main/java/com/github/_1c_syntax/mdclasses/mdo/HTTPServiceMethod.java
@@ -23,6 +23,9 @@ package com.github._1c_syntax.mdclasses.mdo;
 
 import com.github._1c_syntax.mdclasses.mdo.wrapper.DesignerMDO;
 import com.github._1c_syntax.mdclasses.metadata.additional.MDOType;
+import com.github._1c_syntax.mdclasses.unmarshal.converters.MethodHandlerConverter;
+import com.thoughtworks.xstream.annotations.XStreamAlias;
+import com.thoughtworks.xstream.annotations.XStreamConverter;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
@@ -35,15 +38,15 @@ import lombok.ToString;
 public class HTTPServiceMethod extends MDObjectBase {
 
   /**
-   * Обработчик метода http сервиса. Пока строкой
-   * Формат имя метода
-   * Пример ШаблонURLМетод1
+   * Обработчик метода http сервиса
    */
-  private String handler = ""; // TODO сделать классом
+  @XStreamConverter(MethodHandlerConverter.class)
+  @XStreamAlias("handler")
+  private Handler handler;
 
   public HTTPServiceMethod(DesignerMDO designerMDO) {
     super(designerMDO);
-    handler = designerMDO.getProperties().getHandler();
+    handler = new Handler(designerMDO.getProperties().getHandler());
   }
 
   @Override
