+++ src/main/java/com/github/_1c_syntax/mdclasses/mdo/RoleData.java
@@ -39,25 +39,25 @@ import java.util.List;
 @Setter
 public class RoleData {
 
-    /**
-     * Устанавливать права для новых объектов
-     */
-    @XStreamAlias("setForNewObjects")
-    private boolean setForNewObjects = false;
+  /**
+   * Устанавливать права для новых объектов
+   */
+  @XStreamAlias("setForNewObjects")
+  private boolean setForNewObjects = false;
 
-    /**
-     * Устанавливать права для реквизитов и табличных частей по умолчанию
-     */
-    @XStreamAlias("setForAttributesByDefault")
-    private boolean setForAttributesByDefault = false;
+  /**
+   * Устанавливать права для реквизитов и табличных частей по умолчанию
+   */
+  @XStreamAlias("setForAttributesByDefault")
+  private boolean setForAttributesByDefault = false;
 
-    /**
-     * Независимые права подчиненных объектов
-     */
-    @XStreamAlias("independentRightsOfChildObjects")
-    private boolean independentRightsOfChildObjects = false;
+  /**
+   * Независимые права подчиненных объектов
+   */
+  @XStreamAlias("independentRightsOfChildObjects")
+  private boolean independentRightsOfChildObjects = false;
 
-    @XStreamAlias("object")
-    @XStreamImplicit
-    List<ObjectRight> objectRights = Collections.emptyList();
+  @XStreamAlias("object")
+  @XStreamImplicit
+  List<ObjectRight> objectRights = Collections.emptyList();
 }
