+++ src/main/java/com/github/_1c_syntax/mdclasses/mdo/ScheduledJob.java
@@ -23,7 +23,7 @@ package com.github._1c_syntax.mdclasses.mdo;
 
 import com.github._1c_syntax.mdclasses.mdo.wrapper.DesignerMDO;
 import com.github._1c_syntax.mdclasses.metadata.additional.MDOType;
-import com.github._1c_syntax.mdclasses.unmarshal.MethodHandlerConverter;
+import com.github._1c_syntax.mdclasses.unmarshal.converters.MethodHandlerConverter;
 import com.thoughtworks.xstream.annotations.XStreamAlias;
 import com.thoughtworks.xstream.annotations.XStreamConverter;
 import lombok.Data;
