+++ src/main/java/com/github/_1c_syntax/mdclasses/mdo/form/FormAttribute.java
@@ -25,8 +25,8 @@ import com.github._1c_syntax.mdclasses.mdo.form.attribute.DynamicListExtInfo;
 import com.github._1c_syntax.mdclasses.mdo.form.attribute.ExtInfo;
 import com.github._1c_syntax.mdclasses.mdo.wrapper.form.DesignerAttribute;
 import com.github._1c_syntax.mdclasses.mdo.wrapper.form.DesignerColumn;
-import com.github._1c_syntax.mdclasses.unmarshal.ExtInfoConverter;
-import com.github._1c_syntax.mdclasses.unmarshal.ValueTypeConverter;
+import com.github._1c_syntax.mdclasses.unmarshal.converters.ExtInfoConverter;
+import com.github._1c_syntax.mdclasses.unmarshal.converters.ValueTypeConverter;
 import com.thoughtworks.xstream.annotations.XStreamAlias;
 import com.thoughtworks.xstream.annotations.XStreamConverter;
 import com.thoughtworks.xstream.annotations.XStreamImplicit;
