+++ src/main/java/com/github/_1c_syntax/mdclasses/unmarshal/converters/ExtInfoConverter.java
@@ -19,15 +19,20 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with MDClasses.
  */
-package com.github._1c_syntax.mdclasses.unmarshal;
+package com.github._1c_syntax.mdclasses.unmarshal.converters;
 
+import com.github._1c_syntax.mdclasses.mdo.form.attribute.DynamicListExtInfo;
+import com.github._1c_syntax.mdclasses.mdo.form.attribute.ExtInfo;
+import com.github._1c_syntax.mdclasses.unmarshal.XStreamFactory;
 import com.thoughtworks.xstream.converters.Converter;
 import com.thoughtworks.xstream.converters.MarshallingContext;
 import com.thoughtworks.xstream.converters.UnmarshallingContext;
 import com.thoughtworks.xstream.io.HierarchicalStreamReader;
 import com.thoughtworks.xstream.io.HierarchicalStreamWriter;
 
-public class CommandActionConverter implements Converter {
+import java.util.Optional;
+
+public class ExtInfoConverter implements Converter {
   @Override
   public void marshal(Object source, HierarchicalStreamWriter writer, MarshallingContext context) {
     // noop
@@ -35,16 +40,20 @@ public class CommandActionConverter implements Converter {
 
   @Override
   public Object unmarshal(HierarchicalStreamReader reader, UnmarshallingContext context) {
-    reader.moveDown();
-    reader.moveDown();
-    var handler = reader.getValue();
-    reader.moveUp();
-    reader.moveUp();
-    return handler;
+    ExtInfo item;
+    var type = Optional.ofNullable(reader.getAttribute("xsi:type")).orElse(ExtInfo.UNKNOWN);
+    if (type.equals("form:DynamicListExtInfo")) {
+      item = (ExtInfo) context.convertAnother(reader, DynamicListExtInfo.class,
+        XStreamFactory.getReflectionConverter());
+    } else {
+      item = new ExtInfo();
+    }
+    item.setType(type);
+    return item;
   }
 
   @Override
   public boolean canConvert(Class type) {
-    return type == String.class;
+    return type == ExtInfo.class;
   }
 }
