+++ src/main/java/com/github/_1c_syntax/mdclasses/unmarshal/converters/FormItemConverter.java
@@ -19,9 +19,10 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with MDClasses.
  */
-package com.github._1c_syntax.mdclasses.unmarshal;
+package com.github._1c_syntax.mdclasses.unmarshal.converters;
 
-import com.github._1c_syntax.mdclasses.mdo.wrapper.form.DesignerEvent;
+import com.github._1c_syntax.mdclasses.mdo.form.FormItem;
+import com.github._1c_syntax.mdclasses.unmarshal.XStreamFactory;
 import com.thoughtworks.xstream.converters.Converter;
 import com.thoughtworks.xstream.converters.MarshallingContext;
 import com.thoughtworks.xstream.converters.UnmarshallingContext;
@@ -29,9 +30,9 @@ import com.thoughtworks.xstream.io.HierarchicalStreamReader;
 import com.thoughtworks.xstream.io.HierarchicalStreamWriter;
 
 /**
- * Конвертирует событие формы в DesignerEvent
+ * Конвертирует в FormItem элемент формы
  */
-public class FormEventConverter implements Converter {
+public class FormItemConverter implements Converter {
   @Override
   public void marshal(Object source, HierarchicalStreamWriter writer, MarshallingContext context) {
     // noop
@@ -39,12 +40,27 @@ public class FormEventConverter implements Converter {
 
   @Override
   public Object unmarshal(HierarchicalStreamReader reader, UnmarshallingContext context) {
-    var name = reader.getAttribute("name");
-    return new DesignerEvent(name, reader.getValue());
+    // если свойство type отсутствует, тип элемента заполняется из атрибута
+    var nodeName = reader.getNodeName();
+    var type = getItemType(nodeName, reader.getAttribute("xsi:type"));
+    var item = (FormItem) context.convertAnother(reader, FormItem.class,
+      XStreamFactory.getReflectionConverter());
+    if (item.getType().isEmpty()) {
+      item.setType(type);
+    }
+    return item;
   }
 
   @Override
   public boolean canConvert(Class type) {
-    return type == DesignerEvent.class;
+    return type == FormItem.class;
+  }
+
+  private String getItemType(String nodeName, String attribute) {
+    if (nodeName.equals("items")) {
+      return attribute;
+    } else {
+      return nodeName;
+    }
   }
 }
