+++ src/main/java/com/github/_1c_syntax/mdclasses/unmarshal/converters/ValueTypeConverter.java
@@ -19,19 +19,17 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with MDClasses.
  */
-package com.github._1c_syntax.mdclasses.unmarshal;
+package com.github._1c_syntax.mdclasses.unmarshal.converters;
 
-import com.github._1c_syntax.mdclasses.mdo.form.FormItem;
+import com.github._1c_syntax.mdclasses.metadata.additional.ValueType;
+import com.github._1c_syntax.mdclasses.unmarshal.XStreamFactory;
 import com.thoughtworks.xstream.converters.Converter;
 import com.thoughtworks.xstream.converters.MarshallingContext;
 import com.thoughtworks.xstream.converters.UnmarshallingContext;
 import com.thoughtworks.xstream.io.HierarchicalStreamReader;
 import com.thoughtworks.xstream.io.HierarchicalStreamWriter;
 
-/**
- * Конвертирует в FormItem элемент формы
- */
-public class FormItemConverter implements Converter {
+public class ValueTypeConverter implements Converter {
   @Override
   public void marshal(Object source, HierarchicalStreamWriter writer, MarshallingContext context) {
     // noop
@@ -39,27 +37,14 @@ public class FormItemConverter implements Converter {
 
   @Override
   public Object unmarshal(HierarchicalStreamReader reader, UnmarshallingContext context) {
-    // если свойство type отсутствует, тип элемента заполняется из атрибута
-    var nodeName = reader.getNodeName();
-    var type = getItemType(nodeName, reader.getAttribute("xsi:type"));
-    var item = (FormItem) context.convertAnother(reader, FormItem.class,
+    var valueType = (ValueType) context.convertAnother(reader, ValueType.class,
       XStreamFactory.getReflectionConverter());
-    if (item.getType().isEmpty()) {
-      item.setType(type);
-    }
-    return item;
+    return valueType.getTypes();
   }
 
   @Override
   public boolean canConvert(Class type) {
-    return type == FormItem.class;
+    return true;
   }
 
-  private String getItemType(String nodeName, String attribute) {
-    if (nodeName.equals("items")) {
-      return attribute;
-    } else {
-      return nodeName;
-    }
-  }
 }
