+++ src/main/java/com/github/_1c_syntax/mdclasses/unmarshal/converters/package-info.java
@@ -23,4 +23,4 @@
  * Пакет содержит классы для преобразования XML файлов конфигурации в java объекты.
  * Классы конверторы должны обязательно иметь суффикс Converter
  */
-package com.github._1c_syntax.mdclasses.unmarshal;
+package com.github._1c_syntax.mdclasses.unmarshal.converters;
