+++ src/test/java/com/github/_1c_syntax/mdclasses/mdo/EventSubscriptionTest.java
@@ -39,20 +39,20 @@ class EventSubscriptionTest extends AbstractMDOTest {
       "4da21a7b-3d07-4e6d-b91f-7e1c8ddcffcd");
     checkNoChildren(mdo);
     checkNoModules(mdo);
-    assertThat(((EventSubscription) mdo).getHandler())
+    assertThat(((EventSubscription) mdo).getHandler().getMethodPath())
       .isEqualTo("CommonModule.ПростойОбщийМодуль.ПодпискаНаСобытие1ПередЗаписью");
   }
 
   @Override
   @Test
   void testDesigner() {
-    var mdo = getMDObjectEDT("EventSubscriptions/ПодпискаНаСобытие1/ПодпискаНаСобытие1.mdo");
+    var mdo = getMDObjectDesigner("EventSubscriptions/ПодпискаНаСобытие1.xml");
     checkBaseField(mdo, EventSubscription.class, "ПодпискаНаСобытие1",
       "4da21a7b-3d07-4e6d-b91f-7e1c8ddcffcd");
     checkNoChildren(mdo);
     checkNoModules(mdo);
 
-    assertThat(((EventSubscription) mdo).getHandler())
+    assertThat(((EventSubscription) mdo).getHandler().getMethodPath())
       .isEqualTo("CommonModule.ПростойОбщийМодуль.ПодпискаНаСобытие1ПередЗаписью");
   }
 }
